requirements.txt
playwright==1.41.0
playwright-stealth==1.0.6
google-generativeai==0.8.3
python-dotenv==1.0.1